[FEATURE] Write or link central autoload file to typo3/cms package
To let the user define an own vendor dir, we write link the autoload.php
to the TYPO3 package vendor dir folder so that TYPO3 has only one place
to look for such a file.

With that we're bumping the branch to 1.2.x to not interfere with older
TYPO3 versions in this area.