[FEATURE] Make installation paths configurable
Use the "extra" approach from the composer/installers package to make the
installation paths for the TYPO3 source and extensions configurable.

The default behavior/default installation paths are set in the code and
can be change by setting extra -> installer-paths.

This gives more flexibility to the folder structure, e.g. if you don't
want the composer root to be your document root, but want to use a folder
"www-data" inside your distribution (composer root) as your document root.