[TASK] Revert possibility to freely define extension location
The "extension-dir" and "imply-extension-key" options have been removed
to not diverge from composer standard too much and to make it
possible for extensions to ship composer bin scripts. Such scripts need to
know the location of autoload.php, which would be impossible if the extension
directory could be in any place.