[!!!] Remove deprecations and deny installation of typo3/cms
For the 2.0.0 release, which will only work with TYPO3 9.x,
we can remove a lot of code that has been deprecated previously.

Additionally we deny installation of typo3/cms and require people
to us the subtree splitted packages that are available for a while now.

We also now install all extensions exactly the same as a composer library,
by inheriting from the LibraryInstaller. We only change the package types
and the install paths by overriding the according methods.

Since we deny installation of typo3/cms, we can also remove the classes,
that require its presence.