[!!!][BUGFIX] Fix token generation if plugin is symlinked
When installing the plugin from a path directory,
we must ensure the correct directory for the include file
is used for path calculation.

While this is a breaking API change, the API isn't useful
for third party anyway, thus can be treated as internal,
which makes this non breaking for users.