[BUGFIX] Ensure installers only handle supported types
Previously all packages with the type starting
with "typo3-cms-" were handled by the by the
cms extension installer without limit them to
the reserved keys "typo3-cms-extension" for
community extensions and "typo3-cms-framework"
for system extensions.

To ensure that custom handlers can be used here
and avoid unexpected side-effects the extension
the installer now only handles the reserved types.

Requiring the full core package with type "typo3-cms-core"
is not supported anymore. The core installer is now
limited to "typo3-cms-core" to avoid unwanted consequences
due to the fallback to "library" since the package
the type was not explicitly checked before.