[TASK] Streamline CI-Setup
The goal of this patch is to provide a base for multi env
testing of the composer installer. The base matrix now
also includes composer versions 1 and 2.

The missing LICENSE.txt file was added.
The unused .styleci file was removed.

Git is now configured and will exclude ci related files
from shipped packages, in addition, we ensure that
relevant files are checked out with LF line endings
across all platforms.

Editorconfig will help us to ensure consistent line endings
and indentions across the project.

CGL check is now enabled again for ci builds, we are using a
standalone package for php-cs-fixer to avoid version conflicts
with shared dependencies of the installer. The header comment
is now also covered through the CGL to ensure it´s available
in all files.

For linting we are now using the overtrue/phplint package, this also
allows easy linting on non-linux systems and provides better ci fitting
reporting.