adjust downloader return types for composer 2.3 compatibility (#132)
The interface now requires a Promise being returned.
This can simply be adjusted by still running synchronously
and then returning a resolved promise instantly.

The downside is that the new abstract class requires
the return type "?string", which is only possible in php 7.1,
so I had to increase the php version requirement.

I described more considerations in the php file itself.